mcp-server/pyproject.toml
[project]
name = "swim-rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "google-api-python-client>=2.177.0",
    "google-auth[aiohttp]>=2.40.3",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pytest-sugar>=1.0.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "mypy>=1.13.0",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.2",
]

[tool.ruff]
line-length = 80

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
addopts = ["--cov=src", "--cov-config=pyproject.toml"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["import-untyped", "attr-defined"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
